added yaml-parsing from ros_param_server and cleand up code
